Fixed large stack usage in netcdf class

+++ src/data_struct/fit_parameters.cpp
@@ -208,7 +208,7 @@ void Fit_Parameters<T_real>::from_array_d(const std::vector<double> &arr)
 {
     for(auto& itr : _params)
     {
-        if (itr.second.opt_array_index != -1 && itr.second.opt_array_index < (int)arr.size())
+        if (itr.second.opt_array_index != -1 && itr.second.opt_array_index < arr.size())
         {
             itr.second.value = arr[itr.second.opt_array_index];
         }
@@ -224,7 +224,7 @@ void Fit_Parameters<T_real>::from_array(const T_real* arr, size_t arr_size)
     //logit_s<<"\n";
     for(auto& itr : _params)
     {
-        if (itr.second.opt_array_index != -1 && itr.second.opt_array_index < (int)arr_size)
+        if (itr.second.opt_array_index != -1 && itr.second.opt_array_index < arr_size)
         {
             itr.second.value = arr[itr.second.opt_array_index];
         }
